Adds new appdaemon [Mohlsson/ReplyLightsHistory]
This is an AppDaemon App for Home Assistant that is designed to replays your lights behavior when no one is home. There are multiple other approaches out there that try to simulate behavior, which is hard. This approach simply uses your smart devices previous behavior to control their pending behavior.