Adding HA Floorplan
Our main file is called `floorplan.js` in `/dist` as defined in the [requirements](https://hacs.xyz/docs/publish/plugin#repository-structure) of a plugin. 

We'll start using releases, as soon as the docs are 100% ready for the plugin. However, people have now asked for adding it to HACS - and the integration are stable. Therefore, I'm now asking for it to be added :-)