Add PirateWeather as a default repository
Submitting PirateWeather (https://pirateweather.net/) for inclusion in the default repository set. This integration is a copy of the Dark Sky sensor and weather module that does not rely on Dark Sky, providing a fully compatible alternative data source for any integration for users that do not have a Dark Sky API key or after the shutdown date.