flagdays_dk
Flagdays_dk is a integration which is scraping the official flagsdays.

It is possible to add your own flagdays, ex. birthdays or other signifacant events.
You can also specify which flags you own and wish to use.
Furthermore it uses a online API to calculate sunrise and sunset to be use in the flag_up and flag_down times.
If given an offset, it will create some extra attributes with the offset subtracted from the flag_up and flag_down times.

There are boolean switches stating if it is a memorial with special orders regarding the event.