Adds new integration [ITSpecialist111/ai_automation_suggester] (#2793)
* Update integration

Add new integration "ITSpecialist111/ai_automation_suggester"
This is a V1 solution for monitoring entities and suggesting possible automations to the end user on a scheduled basis. This calls OpenAI services to make adaptable suggestions and recommendations.

* Update integration

Added "ITspecialist/AI-automation_suggester" as a new integration that allows for entities to be automatically discovered by OpenAI models and suggests possible automation(s) that could be created from them.

---------

Co-authored-by: Joakim Sørensen <joasoe@gmail.com>